backend/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
gunicorn
